moved rest of skeletons
